added pipes

+++ downloadbitchute.sh
@@ -1,29 +1,79 @@
 #!/bin/bash 
 #./downloadbitchute.sh https://www.bitchute.com/video/xJNBHF80zAI/
 
+paramArr=()
+
 for param in "$@" 
 do
-    Test="${param}";
+    paramArr+=("${param}");
 done
+
+ 
+#echo "${paramArr[1]}"
+#echo ${#paramArr[@]}
+
+OptionKeep=0
+OptionDLVid=1
+OptionWGetArgs="-q --show-progress"
+
+for (( i=0; (i+1)<${#paramArr[@]}; i++))
+do
+  case "${paramArr[i]}" in
+    "-h" | "--help")
+      echo "downloadbitchute.sh { Options } [URL]"
+      echo "Available Options:"
+      echo "--keep: Keep Tmp Files and Skripts"
+      echo "--novid: Skip Video Download"
+      echo "-h/--help: This Help"
+      exit
+    ;;
+    "--keep")
+      OptionKeep=1
+      echo "k param"
+      echo "${paramArr[i+1]}"
+   ;;
+   "--novid")
+     OptionDLVid=0
+   ;;
+    *)
+      #echo "else"
+      #echo "${paramArr[i]}"
+    ;;
+  esac
+  #echo "${paramArr[i]}"
+  #echo "${paramArr[${#paramArr[@]}-1]}"
+done
+
+Test=${paramArr[${#paramArr[@]}-1]}
+
+echo "Keep: $OptionKeep"
+if [ $OptionKeep -lt 1 ]
+then 
+echo "do not keep"
+fi
+echo "Test: $Test"
+#exit
 
 if [ "$Test" = '' ]
 then
 echo "No URL specified"
 exit
-else
-echo "${Test}"  > tmp.txt
 fi
-#gedit tmp.txt
 
-#echo "https://www.bitchute.com/video/xJNBHF80zAI/" > tmp.txt
-sed -i 's/https:\/\/www.bitchute.com\/video//g' tmp.txt
-sed -i 's/?list=subscriptions//g' tmp.txt
-sed -i 's/\///g' tmp.txt
+##echo "https://www.bitchute.com/video/xJNBHF80zAI/" > tmp.txt
+#sed -i 's/https:\/\/www.bitchute.com\/video//g' tmp.txt
+#sed -i 's/?list=subscriptions//g' tmp.txt
+#sed -i 's/\///g' tmp.txt
 URL="https://www.bitchute.com/video/"
-VideoID=$(cat tmp.txt)
+VideoID=$(echo "${Test}" | sed 's/https:\/\/www.bitchute.com\/video//g' | sed 's/?list=subscriptions//g' | sed 's/\///g')
+#VideoID=$(cat tmp.txt)
+
 URL="${URL}${VideoID}/"
 #exit
-rm tmp.txt
+#if [ $OptionKeep -lt 1 ]
+#then 
+#rm tmp.txt
+#fi
 
 #echo $URL
 
@@ -33,7 +83,8 @@ cp dl_*.sh "${VideoID}/"
 cp downloadbitchute-channelinfo.sh "${VideoID}/"
 cd $VideoID
 echo $VideoID > videos.txt
-tmp=$(wget $URL)
+tmp=$(wget $OptionWGetArgs $URL)
+#tmp=$(wget $OptionWGetArgs $URL)
 
 
 #only referedVideoThumb=$(grep "<img class=\"img-responsive lazyload\"" index.html)
@@ -58,21 +109,34 @@ tmp=$(wget $URL)
 ./dl_Title.sh
 ./dl_Description.sh
 #exit
+#echo "download chanel info"
 ./downloadbitchute-channelinfo.sh
 #DownloadText=$(cat tmp.txt)
 DownloadText=$(cat VidURL.txt)
 Channel=$(cat Channel.txt)
 Title=$(cat Title.txt)
 Description=$(cat Description.txt)
+
+#download actual video
+if [ $OptionDLVid -ge 1 ]
+then 
+#tmpclear
+echo "Downloading Video File"
+tmp=$(wget $OptionWGetArgs $DownloadText)
+fi
+
 #rm tmp.txt
 #exit
+if [ $OptionKeep -lt 1 ]
+then 
 rm videos.txt
 rm VidURL.txt
 rm index.html
 rm downloadbitchute-channelinfo.sh
 rm dl_*.sh
+fi
 #echo $DownloadText
 
-tmp=$(wget $DownloadText)
-clear
+#tmp=$(wget $DownloadText)
+#clear
 echo "Done"
